test: fix test sql/func-recreate.test.lua
After commit d4a7459e79, this test does not work properly. The
main purpose of this test at the time of creation was to show that
if an error occurred while creating the SQL function, the user
would be able to see this error. After the commit, the error
disappeared from this test.

The patch modifies the test, indicating that the error will be
shown to the user.

Follow-up #4384