test: make app-tap/init_script produce less diff
When a new option is added, app-tap/init_script
outputs big diff. Because all options are printed with
ordinal indexes. Addition of a new option changes
indexes of all options after the new one.

The patch removes indexes from the output making diff
smaller, when a new option is added.