evio: stop accepting connections on high used descriptors ratio
On connection, iproto thread invokes a callback to accept and start
processing it. In some cases (like tarantool being unresponsive under
high load on start) blind acceptance leads to rlimit exhaustion and
consequent fail with "Too many open files" error. To prevent this we
are not accepting connections when used descriptors ratio it too high.

Closes #3776